requirements.txt

pyzmq
msgpack